backend/requirements.txt
fastapi
uvicorn
pytest
httpx
requests
argon2-cffi
python-jose
caddy-python